Safari-611.1.10/LayoutTests/js/dom/regexp-unicode-handling-expected.txt

Test for proper handling of Unicode RegExps and bug 7445: Gmail puts wrong subject in replies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cy('Re: Moose') is 'Moose'
PASS cy('\u8f6c\u53d1: Moose') is 'Moose'
PASS inlineRe.source is newFromInlineRe.source
PASS inlineRe.source is evalFromInlineRe.source
PASS inlineRe.source is evalInlineRe.source
PASS inlineRe.source is newFromEvalInlineRe.source
PASS inlineRe.source is evalFromEvalInlineRe.source
PASS inlineRe.source is explicitRe.source
PASS inlineRe.source is newFromExplicitRe.source
PASS inlineRe.source is evalFromExplicitRe.source
PASS inlineRe.toString() is newFromInlineRe.toString()
PASS inlineRe.toString() is evalFromInlineRe.toString()
PASS inlineRe.toString() is evalInlineRe.toString()
PASS inlineRe.toString() is newFromEvalInlineRe.toString()
PASS inlineRe.toString() is evalFromEvalInlineRe.toString()
PASS inlineRe.toString() is explicitRe.toString()
PASS inlineRe.toString() is newFromExplicitRe.toString()
PASS inlineRe.toString() is evalFromExplicitRe.toString()
PASS inlineRe.exec(sample)[0] is 'bm⠠p'
PASS evalInlineRe.exec(sample)[0] is 'bm⠠p'
PASS explicitRe.exec(sample)[0] is 'bm⠠p'
PASS binlineRe.source is bnewFromInlineRe.source
PASS binlineRe.source is bevalFromInlineRe.source
PASS binlineRe.source is bevalInlineRe.source
PASS binlineRe.source is bnewFromEvalInlineRe.source
PASS binlineRe.source is bevalFromEvalInlineRe.source
PASS binlineRe.source is bexplicitRe.source
PASS binlineRe.source is bnewFromExplicitRe.source
PASS binlineRe.source is bevalFromExplicitRe.source
PASS binlineRe.toString() is bnewFromInlineRe.toString()
PASS binlineRe.toString() is bevalFromInlineRe.toString()
PASS binlineRe.toString() is bevalInlineRe.toString()
PASS binlineRe.toString() is bnewFromEvalInlineRe.toString()
PASS binlineRe.toString() is bevalFromEvalInlineRe.toString()
PASS binlineRe.toString() is bexplicitRe.toString()
PASS binlineRe.toString() is bnewFromExplicitRe.toString()
PASS binlineRe.toString() is bevalFromExplicitRe.toString()
PASS binlineRe.exec(bsample)[0] is 'bm|p'
PASS bevalInlineRe.exec(bsample)[0] is 'bm|p'
PASS bexplicitRe.exec(bsample)[0] is 'bm|p'
PASS successfullyParsed is true

TEST COMPLETE
